-added debugging to check if paintbox thread halts -adapted step length from step detection in best/multifit to model speed from camera

+++ src/de/uvwxy/footpath2/movement/h263/classifier/MVDClasssifier.java
@@ -1,5 +1,5 @@
 package de.uvwxy.footpath2.movement.h263.classifier;
 
 public interface MVDClasssifier {
-	public int classify(long now_ms, float[][][] mvds);
+	public float classify(long now_ms, float[][][] mvds);
 }
